+++ src/locales/fr-ca.ts
@@ -99,6 +99,9 @@ const frCA: Partial<Locale> = {
   heading: {
     backLinkAriaLabel: () => "Retour",
   },
+  label: {
+    optional: () => "facultatif",
+  },
   link: {
     skipLinkLabel: () => "Passer au contenu principal",
   },
@@ -163,6 +166,24 @@ const frCA: Partial<Locale> = {
   },
   pod: {
     undo: () => "Annuler",
+  },
+  textEditor: {
+    boldAria: () => "Gras",
+    cancelButton: () => "Annuler",
+    cancelButtonAria: () => "Annuler",
+    characterCounter(count: number | string) {
+      return `${typeof count === "number" ? count.toString() : count} caractères restants`;
+    },
+    characterLimit(count: number) {
+      return `Vous avez dépassé de ${count} caractères la limite autorisée`;
+    },
+    contentEditorAria: () => "Éditeur de texte enrichi",
+    italicAria: () => "Italique",
+    orderedListAria: () => "Liste ordonnée",
+    saveButton: () => "Enregistrer",
+    saveButtonAria: () => "Enregistrer",
+    toolbarAriaLabel: () => "Mise en forme",
+    unorderedListAria: () => "Liste non ordonnée",
   },
   search: {
     searchButtonText: () => "Chercher",